refactor: invert extension order of the Overlay class
