+++ packages/overlay/src/OverlayDialog.ts
@@ -14,27 +14,24 @@ import {
     firstFocusableIn,
     firstFocusableSlottedIn,
 } from '@spectrum-web-components/shared/src/first-focusable-in.js';
+import { VirtualTrigger } from './VirtualTrigger.js';
+import {
+    Constructor,
+    OpenableElement,
+    OverlayOpenCloseDetail,
+} from './overlay-types.js';
 import {
     BeforetoggleClosedEvent,
     BeforetoggleOpenEvent,
     guaranteedAllTransitionend,
-    OpenableElement,
-    OverlayBase,
-} from './OverlayBase.js';
-import { VirtualTrigger } from './VirtualTrigger.js';
-import { OverlayOpenCloseDetail } from './overlay-types.js';
-
-type Constructor<T = Record<string, unknown>> = {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    new (...args: any[]): T;
-    prototype: T;
-};
+} from './AbstractOverlay.js';
+import type { AbstractOverlay } from './AbstractOverlay.js';
 
 function nextFrame(): Promise<void> {
     return new Promise((res) => requestAnimationFrame(() => res()));
 }
 
-export function OverlayDialog<T extends Constructor<OverlayBase>>(
+export function OverlayDialog<T extends Constructor<AbstractOverlay>>(
     constructor: T
 ): T & Constructor<ReactiveElement> {
     class OverlayWithDialog extends constructor {
