+++ packages/overlay/src/index.ts
@@ -10,7 +10,6 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 export * from './Overlay.js';
-export * from './OverlayBase.js';
 export * from './OverlayTrigger.js';
 export * from './overlay-types.js';
 export * from './VirtualTrigger.js';
